refactor: add framework layer
