+++ server/Dockerfile
@@ -24,4 +24,4 @@ RUN pip install --no-cache-dir --upgrade -r /app/requirements.txt
 COPY ./src /app/src
 
 EXPOSE 8080
-CMD ["uvicorn", "src.main:app", "--host", "0.0.0.0", "--port", "8080"]
+CMD ["uvicorn", "src.framework.web.main:app", "--host", "0.0.0.0", "--port", "8080"]
